修改算法部分
